platform/x86/amd/hsmp: Move ACPI code to acpi.c
commit 969f915473d8cb0e298234ee313370b6a2da69f3 upstream.

An upcoming change splits HSMP driver into ACPI and platform device
variants. Prepare for the split by moving ACPI related code
to acpi.c from hsmp.c. Common code is kept in hsmp.c.

No logical/functional change.
We still have one driver at this point,
the driver probe will be split in the next patch.

Link: https://lore.kernel.org/r/20241021111428.2676884-6-suma.hegde@amd.com
